Hash Compare program
Added a basic python script that takes an SSH public key and compares it against an MD5/SHA256 hash. If I were to develop this further, I would add features of giving the user ability to encode a public ssh key and store it into the program, implement it into a bigger part of a project or even add a UI using a framework such as Tkinter.